move scraper to db folder
